Add tests to ensure we don't interfere with json gem's output
